+++ lib_xcore_math/src/arch/ref/misc.c
@@ -6,9 +6,9 @@
 #include <math.h>
 
 #include "xmath/xmath.h"
-#include "../../vect/vpu_helper.h"
+#include "vpu_helper.h"
 #include "xmath/xs3/vpu_scalar_ops.h"
-#include "../../vect/vpu_const_vects.h"
+#include "vpu_const_vects.h"
 
 
 void vect_s32_merge_accs(
@@ -75,5 +75,5 @@ float s32_to_f32(
     const int32_t mantissa,
     const exponent_t exp)
 {
-  return ldexpf(mantissa, exp);
+  return ldexpf((float) mantissa, exp);
 }
